ports/psoc6: Removing scan routine and associated files.
